refactor(server): extract colyseus-room module-level code (part of #1563)
Split ~515 lines of module-level code out of apps/server/src/colyseus-room.ts
into a sibling transport/colyseus-room/ package, keeping the
VeilColyseusRoom class intact (Colyseus Room subclasses must be a single
class declaration).

Extracted into transport/colyseus-room/:
  - types.ts       - VeilRoomMetadata, VeilRoomOptions, JoinOptions,
                     RoomRuntimeDependencies, LobbyRoomSummary, and
                     idempotent-action / rate-limit / suspicious-action types
  - constants.ts   - timing constants, MAP_SYNC_CHUNK_SIZE, etc.
  - error-reporting.ts - formatBackgroundTaskDetail,
                         reportBackgroundTaskFailure,
                         reportPersistenceSaveFailure
  - rate-limit.ts  - readWebSocketActionRateLimitConfig,
                     readSuspiciousActionAlertConfig
  - room-utils.ts  - sendMessage, clamp, cloneResourceLedger,
                     rebindWorldStatePlayerId, resolveFocusedMapBounds, etc.
  - runtime.ts     - module-level registries (lobbyRoomSummaries,
                     activeRoomInstances, configuredRoomSnapshotStore),
                     zombie-room cleanup, listLobbyRooms, and
                     configuration helpers
  - index.ts       - barrel re-exports

colyseus-room.ts keeps the VeilColyseusRoom class and re-exports the public
API from the new sibling package so existing callers are unaffected. The
shared module state (e.g. configuredRoomSnapshotStore, nextLobbyRoomOwnerToken)
is accessed via ES module live-bindings so TypeScript narrowing behaves as
before.

No runtime behavior change. Pre-existing lifecycle test failures reproduce
identically on main.

Refs #1563
