+++ apps/server/src/transport/colyseus-room/index.ts
@@ -1,0 +1,50 @@
+export * from "./types";
+export {
+  DEFAULT_GROUP_CHALLENGE_SECRET,
+  EMPTY_ROOM_TTL_MS,
+  MAP_SYNC_CHUNK_PADDING,
+  MAP_SYNC_CHUNK_SIZE,
+  MINOR_PROTECTION_TICK_MS,
+  RECONNECTION_WINDOW_SECONDS,
+  TURN_REMINDER_DISCONNECT_THRESHOLD_MS,
+  TURN_TIMER_TICK_MS,
+  ZOMBIE_ROOM_CLEANUP_INTERVAL_MS
+} from "./constants";
+export {
+  formatBackgroundTaskDetail,
+  reportBackgroundTaskFailure,
+  reportPersistenceSaveFailure
+} from "./error-reporting";
+export {
+  parseEnvNumber,
+  readSuspiciousActionAlertConfig,
+  readWebSocketActionRateLimitConfig
+} from "./rate-limit";
+export {
+  clamp,
+  cloneResourceLedger,
+  compareDefaultPlayerSlotIds,
+  hasBattleSnapshotStore,
+  hasPlayerReportStore,
+  isDefaultPlayerSlotId,
+  readMinimumSupportedClientVersion,
+  rebindWorldStatePlayerId,
+  resolveFocusedMapBounds,
+  sendMessage
+} from "./room-utils";
+export {
+  activeRoomInstances,
+  advanceLobbyRoomOwnerToken,
+  configuredRoomSnapshotStore,
+  configureRoomRuntimeDependencies,
+  configureRoomSnapshotStore,
+  ensureZombieRoomCleanupLoop,
+  getActiveRoomInstances,
+  listLobbyRooms,
+  lobbyRoomOwnerTokens,
+  lobbyRoomSummaries,
+  resetLobbyRoomRegistry,
+  resetRoomRuntimeDependencies,
+  roomRuntimeDependencies,
+  runZombieRoomCleanup
+} from "./runtime";
